+++ methods/src/main/kotlin/methods/algebra/AlgebraPlans.kt
@@ -24,7 +24,10 @@ import methods.constantexpressions.simpleTidyUpSteps
 import methods.general.NormalizationPlans
 import methods.inequations.InequationsPlans
 import methods.polynomials.PolynomialsPlans
-import methods.polynomials.polynomialSimplificationSteps
+import methods.polynomials.addFractionsSteps
+import methods.polynomials.addTermAndFractionSteps
+import methods.polynomials.collectLikeTermsSteps
+import methods.polynomials.simplificationSteps
 import methods.rationalexpressions.RationalExpressionsPlans
 import methods.solvable.computeOverallIntersectionSolution
 
@@ -75,7 +78,7 @@ enum class AlgebraPlans(override val runner: CompositeMethod) : RunnerMethod {
             )
 
             steps {
-                apply(algebraicSimplificationSteps())
+                apply(algebraicSimplificationSteps)
             }
         },
     ),
@@ -141,7 +144,13 @@ enum class AlgebraPlans(override val runner: CompositeMethod) : RunnerMethod {
     ),
 }
 
-fun algebraicSimplificationSteps(addRationalExpressions: Boolean = true): StepsProducer {
+val algebraicSimplificationSteps = algebraicSimplificationSteps(true)
+
+// when solving equations or inequalities we don't want to add fractions
+// e.g. in x/2 + x/3 = 1 we want to multiply through by 6 instead of adding the fractions
+val algebraicSimplificationStepsWithoutFractionAddition = algebraicSimplificationSteps(false)
+
+private fun algebraicSimplificationSteps(addRationalExpressions: Boolean = true): StepsProducer {
     return steps {
         whilePossible { deeply(simpleTidyUpSteps) }
         optionally(NormalizationPlans.NormalizeExpression)
@@ -153,12 +162,23 @@ fun algebraicSimplificationSteps(addRationalExpressions: Boolean = true): StepsP
                     option(RationalExpressionsPlans.SimplifyPowerOfRationalExpression)
                     option(RationalExpressionsPlans.MultiplyRationalExpressions)
                     option(RationalExpressionsPlans.MultiplyRationalExpressionWithNonFractionalFactors)
+                    option(PolynomialsPlans.MultiplyVariablePowers)
+                    option(PolynomialsPlans.MultiplyMonomials)
+                    option(PolynomialsPlans.SimplifyPowerOfNegatedVariable)
+                    option(PolynomialsPlans.SimplifyPowerOfVariablePower)
+                    option(PolynomialsPlans.SimplifyPowerOfMonomial)
+                    option(PolynomialsPlans.SimplifyMonomial)
+
+                    option(simplificationSteps)
+                    option(collectLikeTermsSteps)
+
                     if (addRationalExpressions) {
+                        option(addFractionsSteps)
+                        option(addTermAndFractionSteps)
                         option(RationalExpressionsPlans.AddLikeRationalExpressions)
                         option(RationalExpressionsPlans.AddTermAndRationalExpression)
                         option(RationalExpressionsPlans.AddRationalExpressions)
                     }
-                    option(polynomialSimplificationSteps)
                 }
             }
         }