+++ methods/src/main/kotlin/methods/inequations/InequationsPlans.kt
@@ -23,6 +23,7 @@ import engine.patterns.inequationOf
 import engine.steps.Transformation
 import engine.steps.metadata.MetadataKey
 import engine.steps.metadata.metadata
+import methods.algebra.algebraicSimplificationStepsWithoutFractionAddition
 import methods.constantexpressions.constantSimplificationSteps
 import methods.constantexpressions.simpleTidyUpSteps
 import methods.equations.EquationsPlans
@@ -30,7 +31,6 @@ import methods.equations.EquationsRules
 import methods.factor.FactorPlans
 import methods.general.NormalizationPlans
 import methods.polynomials.PolynomialsPlans
-import methods.polynomials.polynomialSimplificationSteps
 import methods.solvable.SolvablePlans
 import methods.solvable.SolvableRules
 import methods.solvable.evaluateBothSidesNumerically
@@ -46,7 +46,7 @@ enum class InequationsPlans(override val runner: CompositeMethod) : RunnerMethod
                 optionally(NormalizationPlans.NormalizeExpression)
                 whilePossible(EquationsRules.EliminateConstantFactorOfLhsWithZeroRhs)
                 whilePossible(SolvableRules.CancelCommonTermsOnBothSides)
-                whilePossible(polynomialSimplificationSteps)
+                whilePossible(algebraicSimplificationStepsWithoutFractionAddition)
             }
         },
     ),