+++ CHANGELOG.md
@@ -31,6 +31,8 @@ contents to a new release section (e.g. _[0.x]_).
 - SDK: add support for specials symbols i.e. `%, pi, exponentialE, iota` (PLUT-822)
 - Added support for parametric quadratic equations (where coefficents are not constant) (PLUT-807)
 - Added ability to simplify a square root with a square factor (e.g. `sqrt[4x - 8]`) (PLUT-808)
+- Added support for simplifying the sum of non-constant fractions, such as `[x + 1 / 2] + [3x / 5]` (not applied when
+  solving equations) (PLUT-680)
 
 ### Fixed